modify admin, mypage directory structure
