--- src/main/java/com/booktalk_be/domain/admin/service/AdminServiceImpl.java
@@ -1,4 +1,0 @@
-package com.booktalk_be.domain.admin.service;
-
-public class AdminServiceImpl implements AdminService {
-}
