+++ src/main/java/com/booktalk_be/domain/member/auth/command/CreateMemberCommand.java
@@ -1,4 +1,4 @@
-package com.booktalk_be.domain.member.join.command;
+package com.booktalk_be.domain.member.auth.command;
 
 import com.booktalk_be.domain.member.auth.model.entity.AuthenticateType;
 import com.booktalk_be.domain.member.auth.model.entity.AuthorityType;
@@ -11,7 +11,16 @@ import java.time.LocalDate;
 
 @Getter
 @NoArgsConstructor
-public class ModifyMemberCommand {
+public class CreateMemberCommand {
+
+    @NotNull
+    private String email;
+
+    @NotNull
+    private String name;
+
+    @NotNull
+    private AuthenticateType authType;
 
     @NotNull
     private String password;
@@ -24,11 +33,16 @@ public class ModifyMemberCommand {
 
     private LocalDate birth;
 
+    private AuthorityType authority;
+
     @NotNull
     private Boolean delYn = false;
 
     public Member toEntity() {
         return Member.builder()
+                .email(this.email)
+                .name(this.name)
+                .authType(this.authType)
                 .password(this.password)
                 .phoneNumber(this.phoneNumber)
                 .address(this.address)
