+++ src/main/java/com/booktalk_be/domain/member/auth/command/ModifyMemberCommand.java
@@ -1,10 +1,6 @@
-package com.booktalk_be.domain.member.join.command;
+package com.booktalk_be.domain.member.auth.command;
 
-import com.booktalk_be.domain.board.model.entity.Board;
-import com.booktalk_be.domain.member.auth.model.entity.AuthenticateType;
-import com.booktalk_be.domain.member.auth.model.entity.AuthorityType;
 import com.booktalk_be.domain.member.auth.model.entity.Member;
-import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,16 +9,7 @@ import java.time.LocalDate;
 
 @Getter
 @NoArgsConstructor
-public class CreateMemberCommand {
-
-    @NotNull
-    private String email;
-
-    @NotNull
-    private String name;
-
-    @NotNull
-    private AuthenticateType authType;
+public class ModifyMemberCommand {
 
     @NotNull
     private String password;
@@ -35,16 +22,11 @@ public class CreateMemberCommand {
 
     private LocalDate birth;
 
-    private AuthorityType authority;
-
     @NotNull
     private Boolean delYn = false;
 
     public Member toEntity() {
         return Member.builder()
-                .email(this.email)
-                .name(this.name)
-                .authType(this.authType)
                 .password(this.password)
                 .phoneNumber(this.phoneNumber)
                 .address(this.address)
